適応度の計算について、なるべく線形になるように変更。

+++ src/main/java/com/kurukurupapa/pff/domain/JobSkill.java
@@ -52,19 +52,19 @@ public class JobSkill {
 	 *            ジョブスキル適用前の回復値
 	 * @return 回復値
 	 */
-	public int calcRecoveryMagicDamage(int recoveryDamage) {
-		int damage = recoveryDamage;
+	public float calcRecoveryMagicDamage(float recoveryDamage) {
+		float damage = recoveryDamage;
 		if (mName.equals("オーラ")) {
 			// オーラ
 			// 自身の白魔法アビリティ効果が上昇
 			// 回復量1.8倍 他の白魔法にも効果あり
-			damage = (int) (recoveryDamage * 1.8f);
+			damage = recoveryDamage * 1.8f;
 		}
 		return damage;
 	}
 
-	public int getFureaAttackDamage(int intelligence, float magicAttack) {
-		return (int) ((intelligence + 100) * magicAttack * Mement.CHIE_BLACK_RATE);
+	public float getFureaAttackDamage(int intelligence, float magicAttack) {
+		return (intelligence + 100f) * magicAttack * Mement.CHIE_BLACK_RATE;
 	}
 
 }
